Notification style change and file re-structuring
Notification style change and file re-structuring for ease of future use.
