+++ package.json
@@ -5,7 +5,8 @@
     "systeminformation": "^5.25.11"
   },
   "scripts": {
-    "dev": "nodemon app --port 3001"
+    "start": "npm i && npm update && node server/app --port 3001",
+    "dev": "nodemon server/app --port 3001"
   },
   "devDependencies": {
     "nodemon": "^3.1.9"
