+++ src/index.html
@@ -94,4 +94,7 @@
 </div>
 </body>
 <script src="static/js/main.js"></script>
+<script src="static/js/utils.js"></script>
+<script src="static/js/notification.js"></script>
+<script src="static/js/loading.js"></script>
 </html>
